backend/requirements/base.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
dj-database-url==2.2.0
pydantic==2.9.2
cryptography==43.0.1
requests==2.32.3
openai>=1.54.0
beautifulsoup4==4.12.3
lxml==5.3.0

